Fixed CodeQL.

+++ src/main/java/g0701_0800/s0793_preimage_size_of_factorial_zeroes_function/Solution.java
@@ -21,11 +21,11 @@ public class Solution {
     }
 
     private int countZeros(long n) {
-        int rst = 0;
+        long rst = 0;
         while (n > 0) {
             rst += n / 5;
             n /= 5;
         }
-        return rst;
+        return (int) rst;
     }
 }
